_ref/auto-rag-optimizer/pyproject.toml
[project]
name = "auto-rag-optimizer"
version = "0.1.0"
description = "Autonomous RAG configuration optimizer inspired by Karpathy's autoresearch"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.8.0",
    "ragas>=0.2.0",
    "openai>=1.50.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "unstructured>=0.15.0",
    "tiktoken>=0.7.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
